feat: 로또 구입 시 수동으로 입력하는 기능 구현

+++ src/main/java/camp/nextstep/edu/level1/stringCaluator/calculator/StringCalculator.java
@@ -10,7 +10,7 @@ public class StringCalculator {
         this.positiveNumbers = new PositiveNumbers(StringParser.parseAndSplit(input));
     }
 
-    public int sum() {
+    public long sum() {
         return this.positiveNumbers.sum();
     }
 }
