+++ src/test/java/camp/nextstep/edu/common/PositiveNumbersTest.java
@@ -11,20 +11,20 @@ class PositiveNumbersTest {
 
     @Test
     void 문자열_양수와_숫자_양수로_양수_리스트_객체를_생성하면_정상_생성된다() {
-        List<Integer> numbers = Arrays.asList(1, 2, 3, 4, 5);
+        List<Long> numbers = Arrays.asList(1L, 2L, 3L, 4L, 5L);
         List<String> stringNumbers = Arrays.asList("1", "2", "3", "4", "5");
 
         PositiveNumbers positiveNumbersByNumber = new PositiveNumbers(numbers);
         PositiveNumbers positiveNumbersByStringNumbers = new PositiveNumbers(stringNumbers);
 
-        assertThat(positiveNumbersByNumber.sum()).isEqualTo(numbers.stream().reduce(0, Integer::sum));
+        assertThat(positiveNumbersByNumber.sum()).isEqualTo(numbers.stream().reduce(0L, Long::sum));
         assertThat(positiveNumbersByStringNumbers.sum())
                 .isEqualTo(stringNumbers.stream().mapToInt(Integer::parseInt).sum());
     }
 
     @Test
     void 음수가_포함된_숫자_리스트로_양수_리스트를_생성하면_예외가_발생한다() {
-        List<Integer> numbers = Arrays.asList(1, 2, -5, 3, -8);
+        List<Long> numbers = Arrays.asList(1L, 2L, -5L, 3L, -8L);
 
         assertThatThrownBy(() -> new PositiveNumbers(numbers)).isInstanceOf(RuntimeException.class);
     }
